Modify tinyubsan from github.com/Abb1x/tinyubsan
I have done this work before but using this "Unlicense" version
saves me from having to figure out all the functions again.

I changed the callback signature to not be variable length
and to also contain the return address so you can find
where the offense occurred using addr2line depending on your
target.

You could also define a breakpoint in the callback for easy use
with a debugger.

Future steps are to link it up with my Makefile template and
figure out how to use objcopy to reduce the size of the ubsan
stuff until GCC gets a -fsanitize-undefined-minimal-runtime like
Clang has.